requirements.txt
streamlit
numpy
pandas
tensorflow
scikit-learn
ultralytics
opencv-python-headless
pyserial
librosa
soundfile

